[Oztechan/CCC#2527] Move serialised models to :client:core:remoteconfig
