+++ client/configservice/ad/src/commonMain/kotlin/com/oztechan/ccc/client/configservice/ad/mapper/AdConfigMapper.kt
@@ -1,7 +1,7 @@
 package com.oztechan.ccc.client.configservice.ad.mapper
 
-import com.oztechan.ccc.client.configservice.ad.AdConfig as AdConfigRCModel
 import com.oztechan.ccc.client.configservice.ad.model.AdConfig as AdConfigModel
+import com.oztechan.ccc.client.core.remoteconfig.model.AdConfig as AdConfigRCModel
 
 internal fun AdConfigRCModel.toAdConfigModel() = AdConfigModel(
     bannerAdSessionCount = bannerAdSessionCount,
