+++ client/configservice/ad/src/commonTest/kotlin/com/oztechan/ccc/client/configservice/ad/mapper/AdConfigMapperTest.kt
@@ -2,7 +2,7 @@ package com.oztechan.ccc.client.configservice.ad.mapper
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
-import com.oztechan.ccc.client.configservice.ad.AdConfig as AdConfigRCModel
+import com.oztechan.ccc.client.core.remoteconfig.model.AdConfig as AdConfigRCModel
 
 class AdConfigMapperTest {
