+++ client/configservice/review/src/commonMain/kotlin/com/oztechan/ccc/client/configservice/review/mapper/ReviewConfigMapper.kt
@@ -1,7 +1,7 @@
 package com.oztechan.ccc.client.configservice.review.mapper
 
-import com.oztechan.ccc.client.configservice.review.ReviewConfig as ReviewConfigRCModel
 import com.oztechan.ccc.client.configservice.review.model.ReviewConfig as ReviewConfigModel
+import com.oztechan.ccc.client.core.remoteconfig.model.ReviewConfig as ReviewConfigRCModel
 
 internal fun ReviewConfigRCModel.toReviewConfigModel() = ReviewConfigModel(
     appReviewSessionCount = appReviewSessionCount,
