+++ client/configservice/update/src/commonMain/kotlin/com/oztechan/ccc/client/configservice/update/mapper/UpdateConfigMapper.kt
@@ -1,7 +1,7 @@
 package com.oztechan.ccc.client.configservice.update.mapper
 
-import com.oztechan.ccc.client.configservice.update.UpdateConfig as UpdateConfigRCModel
 import com.oztechan.ccc.client.configservice.update.model.UpdateConfig as UpdateConfigModel
+import com.oztechan.ccc.client.core.remoteconfig.model.UpdateConfig as UpdateConfigRCModel
 
 internal fun UpdateConfigRCModel.toUpdateConfigModel() = UpdateConfigModel(
     updateLatestVersion = updateLatestVersion,
