+++ client/core/remoteconfig/src/commonMain/kotlin/com/oztechan/ccc/client/core/remoteconfig/model/AdConfig.kt
@@ -1,10 +1,10 @@
-package com.oztechan.ccc.client.configservice.ad
+package com.oztechan.ccc.client.core.remoteconfig.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-internal data class AdConfig(
+data class AdConfig(
     @SerialName("banner_ad_session_count") val bannerAdSessionCount: Int = 2,
     @SerialName("interstitial_ad_session_count") val interstitialAdSessionCount: Int = 5,
     @SerialName("interstitial_ad_initial_delay") val interstitialAdInitialDelay: Long = 60000,
