+++ client/core/remoteconfig/src/commonMain/kotlin/com/oztechan/ccc/client/core/remoteconfig/model/ReviewConfig.kt
@@ -1,10 +1,10 @@
-package com.oztechan.ccc.client.configservice.review
+package com.oztechan.ccc.client.core.remoteconfig.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-internal data class ReviewConfig(
+data class ReviewConfig(
     @SerialName("app_review_session_count") val appReviewSessionCount: Int = 3,
     @SerialName("app_review_dialog_delay") val appReviewDialogDelay: Long = 20000
 )
