+++ client/core/remoteconfig/src/commonMain/kotlin/com/oztechan/ccc/client/core/remoteconfig/model/UpdateConfig.kt
@@ -1,10 +1,10 @@
-package com.oztechan.ccc.client.configservice.update
+package com.oztechan.ccc.client.core.remoteconfig.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-internal data class UpdateConfig(
+data class UpdateConfig(
     @SerialName("update_latest_version") val updateLatestVersion: Int = 0,
     @SerialName("update_force_version") val updateForceVersion: Int = 0
 )
